Hybrid Multiple Cryptography Version-2
This scheme implements an approach
that seeks to incorporate both Hybrid cryptography and Multiple cryptography as a way to
provide an enhanced layer of security by combining both of their strengths while also
mitigating the drawbacks that are present in each individual approach. The process begins
with the creation of a symmetric key as well as an asymmetric key pair consisting of a public
and a private key. Initially, the message is encrypted using the symmetric key encryption
technique with the symmetric key that can encrypt the message for n number of times. The
resulting encrypted text and symmetric key are encrypted with the asymmetric public key
using the asymmetric key encryption algorithm. The decryption process, on the other hand, is
a reversal of the encryption process.Thus the scheme is a reliable and effective data
encryption system that can offer greater data protection than conventional encryption
approaches.